requirements.txt
feedgen
Flask
Flask-Caching
Flask-Limiter
gunicorn
python-dotenv
redis
requests